fix recursive json and bad api impls

+++ server/src/main/java/com/mera/borisgk98/autoshowroom/server/repositories/AutoRepository.java
@@ -1,9 +1,9 @@
 package com.mera.borisgk98.autoshowroom.server.repositories;
 
-import com.mera.borisgk98.autoshowroom.server.models1.Auto;
+import com.mera.borisgk98.autoshowroom.server.models.Auto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AutoRepository extends JpaRepository<Auto, Long> {
+public interface AutoRepository extends JpaRepository<Auto, Integer> {
 }
