+++ server/src/main/java/com/mera/borisgk98/autoshowroom/server/services/AutoService.java
@@ -1,6 +1,6 @@
 package com.mera.borisgk98.autoshowroom.server.services;
 
-import com.mera.borisgk98.autoshowroom.server.models1.Auto;
+import com.mera.borisgk98.autoshowroom.server.models.Auto;
 
-public interface AutoService extends CrudService<Auto, Long>{
+public interface AutoService extends CrudService<Auto, Integer>{
 }
