+++ server/src/main/java/com/mera/borisgk98/autoshowroom/server/services/AutoServiceImpl.java
@@ -1,13 +1,13 @@
 package com.mera.borisgk98.autoshowroom.server.services;
 
-import com.mera.borisgk98.autoshowroom.server.models1.Auto;
+import com.mera.borisgk98.autoshowroom.server.models.Auto;
 import com.mera.borisgk98.autoshowroom.server.repositories.AutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class AutoServiceImpl
-        extends AbstractCrudService<Auto, Long>
+        extends AbstractCrudService<Auto, Integer>
         implements AutoService {
 
     @Autowired
