+++ server/src/main/java/com/mera/borisgk98/autoshowroom/server/services/OrderServiceImpl.java
@@ -1,10 +1,10 @@
 package com.mera.borisgk98.autoshowroom.server.services;
 
-import com.mera.borisgk98.autoshowroom.server.models1.Order;
+import com.mera.borisgk98.autoshowroom.server.models.Order;
 import com.mera.borisgk98.autoshowroom.server.repositories.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 
-public class OrderServiceImpl extends AbstractCrudService<Order, Long> {
+public class OrderServiceImpl extends AbstractCrudService<Order, Integer> {
 
     @Autowired
     protected OrderRepository repository;
